Move LogoLink to Components
Move LogoLink to Components/Common/Links.
